Added Database
Added database, it is working and you can use SQLiteStudio to check if your queries in the program is changing the database on the device/emulator that you are using. There are views like home/gallery/slideshow which are not part of the program and just part of the default items in the activity I started with. I've included links with some helpful tips to access the database and write data to the database. 'Create Library' does not do anything except update the database,  ex, no navigation after create is clicked but this is so that we can start working on the database functionality.